Fixed movement using old input system
For some reason the new input system doesn't work on my project, I tried troubleshooting this for hours. Unity detects my keypresses but I can't get the message to send to OnMove() (i put debug logs in there). I used the old input system for this project for now. It's just really weird because the new one has worked in the past for me but it just refuses. I don't really know.